Moved short circuiting tests to JVM platform
