+++ tests/jvm/src/test/scala/cats/effect/IOPlatformSpecification.scala
@@ -27,6 +27,7 @@ import cats.effect.unsafe.{
   WorkStealingThreadPool
 }
 import cats.effect.unsafe.metrics.PollerMetrics
+import cats.effect.syntax.all._
 import cats.syntax.all._
 
 import org.scalacheck.Prop.forAll
@@ -809,6 +810,26 @@ trait IOPlatformSpecification extends DetectPlatform { self: BaseSpec with Scala
         } finally {
           runtime.shutdown()
           schedShut()
+        }
+      }
+
+      "parTraverseN" >> {
+        "short-circuit on error" in real {
+          case object TestException extends RuntimeException
+          val target = 0.until(100000).toList
+          val test = target.parTraverseN(2)(_ => IO.raiseError(TestException))
+
+          test.attempt.as(ok).timeoutTo(500.millis, IO(false must beTrue))
+        }
+      }
+
+      "parTraverseN_" >> {
+        "short-circuit on error" in real {
+          case object TestException extends RuntimeException
+          val target = 0.until(100000).toList
+          val test = target.parTraverseN_(2)(_ => IO.raiseError(TestException))
+
+          test.attempt.as(ok).timeoutTo(500.millis, IO(false must beTrue))
         }
       }
